embedding domain model tidy up
