+++ opsmate/knowledgestore/models.py
@@ -8,6 +8,9 @@ from typing import List
 import uuid
 from enum import Enum
 from datetime import datetime
+from abc import ABC, abstractmethod
+from functools import cache
+from openai import AsyncOpenAI
 
 registry = get_registry()
 
@@ -44,6 +47,55 @@ class KnowledgeStore(LanceModel):
     created_at: datetime = Field(
         description="The created at date of the knowledge", default_factory=datetime.now
     )
+
+
+class EmbeddingClient(ABC):
+    @abstractmethod
+    async def embed(self, query: str) -> List[float]:
+        pass
+
+
+class OpenAIEmbeddingClient(EmbeddingClient):
+    def __init__(self, model_name: str = config.embedding_model_name):
+        self.model_name = model_name
+
+    async def embed(self, query: str) -> List[float]:
+        response = await self.embed_client().embeddings.create(
+            input=query, model=self.model_name
+        )
+        return response.data[0].embedding
+
+    @cache
+    def embed_client(self):
+        return AsyncOpenAI()
+
+
+class SentenceTransformersEmbeddingClient(EmbeddingClient):
+    def __init__(self, model_name: str = config.embedding_model_name):
+        self.model_name = model_name
+
+    async def embed(self, query: str) -> List[float]:
+        return self.model().encode(query)
+
+    @cache
+    def model(self):
+        import sentence_transformers
+
+        return sentence_transformers.SentenceTransformer(self.model_name)
+
+
+def init_embedding_client():
+    if config.embedding_registry_name == "openai":
+        return OpenAIEmbeddingClient()
+    elif config.embedding_registry_name == "sentence-transformers":
+        return SentenceTransformersEmbeddingClient()
+    else:
+        raise ValueError(
+            f"Unsupported embedding client: {config.embedding_registry_name}"
+        )
+
+
+embedding_client = init_embedding_client()
 
 
 async def aconn():
